refactor: deposit example
