+++ examples/appConfig.ts
@@ -1,0 +1,7 @@
+export const config = {
+  blockchainWsUrl: process.env.APP_BLOCKCHAIN_WEBSOCKET_URL,
+  clientApiUrl: process.env.APP_CLIENT_API_URL,
+  ethereumPrivateKey: process.env.APP_ETH_PRIVATE_KEY,
+  nightfallMnemonic: process.env.APP_NIGHTFALL_MNEMONIC,
+  tokenAddress: process.env.APP_TOKEN_ADDRESS_1,
+};
